Adds Final Evaluation Feature
After the attendance for the final 8th day is submitted, the Final Evaluation button is unlocked. This will allow the user to enter a comment and a score adjustment for each active trainee. Once submitted, these final evals will be stored in the comment database with the "incidentType" of "Final Evaluation".